backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
celery[redis]==5.3.6
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pyyaml==6.0.1
httpx==0.27.0
msal==1.28.0
aiosmtplib==3.0.1
jinja2==3.1.4
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
anyio==4.3.0
ruff==0.4.4
mypy==1.10.0
